% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{generate.table.data}
\alias{generate.table.data}
\title{Generates HPV Information Centre referenced table structure.}
\usage{
generate.table.data(data.dict, selected.columns = NULL,
  column.names = NULL, row.id.column = NULL, format = "latex",
  collapse.refs = TRUE)
}
\value{
List with three components: data, columns and references,
where data is the original dataframe with the necessary reference markers,
columns is the list with the table columns as will be displayed and
references is a nested structure: for each type (sources, notes, ...)
there is a list of dictionaries with each ('marker' key) and associated
reference ('text' key).
}
\description{
Generates a new dataframe with the markers corresponding to the defined
references (sources, notes, ...), alongside a list of the markers' meaning.
}
\details{
@param data.dict List returned by the IC data fetcher
 @param selected.columns Column names to be selected
 from the original dataframe
 @param column.names List with the column names of the selected
 dataframe columns
 @param row.id.column Column name that will contain the marks for row
 references
 @param format Format of the returned dataframe
 @param collapse.refs Whether markers should be collapsed when
 appropriate (e.g. all cells of a column to the column header)
}
\examples{
# TODO

}
