% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{method.markers}
\alias{method.markers}
\title{Generates iterator for method markers}
\usage{
method.markers()
}
\value{
Iterator for method markers
}
\description{
Generates an iterator for successive method markers to be used as footnotes.
Currently these markers are formatted as LaTeX.
}
\examples{
markers <- method.markers()
nextElem(markers)
# [1] '\\\\alpha'
nextElem(markers)
# [1] '\\\\beta'

}
