% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{note.markers}
\alias{note.markers}
\title{Generates iterator for note markers}
\usage{
note.markers()
}
\value{
Iterator for note markers
}
\description{
Generates an iterator for successive note markers to be used as footnotes.
}
\examples{
markers <- note.markers()
nextElem(markers)
# [1] 'a'
nextElem(markers)
# [1] 'b'

}
