% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{source.markers}
\alias{source.markers}
\title{Generates iterator for source markers}
\usage{
source.markers()
}
\value{
Iterator for source markers
}
\description{
Generates an iterator for successive source markers to be used as footnotes.
}
\examples{
markers <- source.markers()
nextElem(markers)
# [1] '1'
nextElem(markers)
# [1] '2'

}
