% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{year.markers}
\alias{year.markers}
\title{Generates iterator for year markers}
\usage{
year.markers()
}
\value{
Iterator for year markers
}
\description{
Generates an iterator for successive year markers to be used as footnotes.
Currently these markers are formatted as LaTeX.
}
\examples{
markers <- year.markers()
nextElem(markers)
# [1] '\\\\Diamond'
nextElem(markers)
# [1] '\\\\triangle'

}
